+++ stringmatch/args.py
@@ -1,12 +1,12 @@
 from typing import Optional, TypedDict
 
-from stringmatch.scorer import _Scorer
+from stringmatch.scorer import BaseScorer
 
 
 class RatioKwargs(TypedDict, total=False):
     """Contains the kwargs for the Ratio class, for type hinting purposes."""
 
-    scorer: type[_Scorer]
+    scorer: type[BaseScorer]
     score: int
     limit: Optional[int]
     latinise: bool